Move the UIOutput class into a new "ui" directory
